refactor gorb into own files, storage on its own
