+++ app/payload/control_freak/flight-software/src/storage.h
@@ -1,0 +1,2 @@
+// entry point for storage thread (will in the future be setup to handle more data)
+int storage_thread_entry(void *, void *, void *);
